Commiting editor's suggested changes
